+++ flows/landing_zone_delete.py
@@ -12,6 +12,9 @@ class Flow(BaseLinearFlow):
     """Flow for deleting a landing zone from a project and a user in iRODS"""
 
     def validate(self):
+        self.supported_modes = [
+            'sync',
+            'async']
         self.required_fields = [
             'zone_title',
             'zone_uuid',
@@ -31,9 +34,31 @@ class Flow(BaseLinearFlow):
             assay_path=self.flow_data['assay_path'],
             zone_title=self.flow_data['zone_title'])
 
-        ##############
-        # iRODS Tasks
-        ##############
+        ########
+        # Tasks
+        ########
+
+        # If async, set up task to set landing zone status to failed
+        if self.request_mode == 'async':
+            self.add_task(
+                omics_tasks.RevertLandingZoneFailTask(
+                    name='Set landing zone status to FAILED on revert',
+                    omics_api=self.omics_api,
+                    project_uuid=self.project_uuid,
+                    inject={
+                        'zone_uuid': self.flow_data['zone_uuid'],
+                        'info_prefix': 'Running asynchronous job failed'}))
+
+        # Set zone status to DELETING
+        self.add_task(
+            omics_tasks.SetLandingZoneStatusTask(
+                name='Set landing zone status to DELETING',
+                omics_api=self.omics_api,
+                project_uuid=self.project_uuid,
+                inject={
+                    'zone_uuid': self.flow_data['zone_uuid'],
+                    'status': 'DELETING',
+                    'status_info': 'Deleting landing zone'}))
 
         self.add_task(
             irods_tasks.RemoveCollectionTask(
@@ -46,10 +71,13 @@ class Flow(BaseLinearFlow):
         # Omics Data Access Tasks
         ##########################
 
+        # Set zone status to DELETING
         self.add_task(
-            omics_tasks.RemoveLandingZoneTask(
-                name='Remove the landing zone from the Omics database',
+            omics_tasks.SetLandingZoneStatusTask(
+                name='Set landing zone status to DELETED',
                 omics_api=self.omics_api,
                 project_uuid=self.project_uuid,
                 inject={
-                    'zone_uuid': self.flow_data['zone_uuid']}))
+                    'zone_uuid': self.flow_data['zone_uuid'],
+                    'status': 'DELETED',
+                    'status_info': 'Landing zone deleted'}))